Pull out accordion styling to global scale
