--- packages/oscal-react-library/src/components/StyedAccordion.tsx
@@ -1,37 +1,0 @@
-import ArrowForwardSharp from "@mui/icons-material/ArrowForwardIosSharp";
-import MuiAccordion from "@mui/material/Accordion";
-import MuiAccordionDetails from "@mui/material/AccordionDetails";
-import MuiAccordionSummary from "@mui/material/AccordionSummary";
-import { styled } from "@mui/material/styles";
-import React from "react";
-
-export const Accordion = styled((props: any) => (
-  <MuiAccordion disableGutters elevation={0} square {...props} />
-))(({ theme }) => ({
-  border: `1px solid ${theme.palette.divider}`,
-  "&:not(:last-child)": {
-    borderBottom: 0,
-  },
-  "&:before": {
-    display: "none",
-  },
-})) as typeof MuiAccordion;
-
-export const AccordionSummary = styled((props) => (
-  <MuiAccordionSummary expandIcon={<ArrowForwardSharp sx={{ fontSize: "0.9rem" }} />} {...props} />
-))(({ theme }) => ({
-  backgroundColor:
-    theme.palette.mode === "dark" ? "rgba(255, 255, 255, .05)" : "rgba(0, 0, 0, .03)",
-  flexDirection: "row-reverse",
-  "& .MuiAccordionSummary-expandIconWrapper.Mui-expanded": {
-    transform: "rotate(90deg)",
-  },
-  "& .MuiAccordionSummary-content": {
-    marginLeft: theme.spacing(1),
-  },
-})) as typeof MuiAccordionSummary;
-
-export const AccordionDetails = styled(MuiAccordionDetails)(({ theme }) => ({
-  padding: theme.spacing(2),
-  borderTop: "1px solid rgba(0, 0, 0, .125)",
-})) as typeof MuiAccordionDetails;
